training-orchestrator/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
sqlalchemy
pytest>=8.3.0
pytest-asyncio>=0.23.2
pytest-html>=3.1.0
pytest-cov>=3.0.0
pytest-mock==3.12.0
httpx>=0.23.0
requests>=2.31.0
pydantic
